lambda/requirements.txt
# TrIAge Lambda Dependencies
# Optimized for AWS Lambda

# Core dependencies from main project
requests>=2.31.0
markdown>=3.5.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# JWT authentication
PyJWT>=2.8.0

# AWS SDK
boto3>=1.34.0
